requirements.txt
djangorestframework
markdown
django-filter
Pillow
django-cors-headers
drf_yasg